chore: Move drawer UI into its own component
